task-3 service handlers in separate files
