+++ product-service/handler.js
@@ -1,29 +1,8 @@
 'use strict';
-const fs = require('fs').promises;
-const path = require('path');
 
-const { PRODUCTS_MOCK_FILENAME } = require('./constants.js')
+const getProductsList = require('./getProductsList.js');
+const getProductById = require('./getProductById.js');
 
-const readMockFile = () => fs.readFile(path.resolve(__dirname, `./${PRODUCTS_MOCK_FILENAME}`));
 
-module.exports.getProductsList = async () => {
-  const productsRaw = await readMockFile();
-  const products = JSON.parse(productsRaw.toString('utf-8'));
-
-  return {
-    statusCode: 200,
-    body: JSON.stringify(products),
-  };
-};
-
-module.exports.getProductById = async (event) => {
-  const {id} = event.pathParameters;
-  const productsRaw = await readMockFile();
-  const products = JSON.parse(productsRaw);
-  const product = products.find((item)=> item.id === id);
-
-  return {
-    statusCode: 200,
-    body: JSON.stringify(product),
-  };
-};
+module.exports.getProductsList = getProductsList;
+module.exports.getProductById = getProductById;
