+++ product-service/utils.js
@@ -1,0 +1,8 @@
+const fs = require('fs').promises;
+const path = require('path');
+
+const { PRODUCTS_MOCK_FILENAME } = require('./constants.js')
+
+const readMockFile = () => fs.readFile(path.resolve(__dirname, `./${PRODUCTS_MOCK_FILENAME}`));
+
+module.exports.readMockFile = readMockFile;
